feat(workspace): 将绑定飞书按钮从Bot详情页移至智能体列表页
- WorkspaceAgentPanel 新增「绑定飞书」按钮，作为创建智能体的入口
- 工作空间下无智能体时，可直接点击绑定飞书来添加
- BotDetailPage 移除绑定飞书功能（绑定是创建行为，不应在详情页）
- BotDetailPage 聊天绑定 Tab 改为纯展示模式
